requirements.txt
pytorch
jupyterlab
natsort
opencv-python
scikit-image~=0.19.2
tensorboardX~=2.5.1
tb-nightly
numpy~=1.16.6
# torch~=1.11.0
pillow~=9.0.1
torchvision~=0.12.0
scipy~=1.7.3
ipywidgets~=7.6.5
matplotlib~=3.4.3
scikit-learn~=1.0.2
torchviz~=0.0.2
piqa~=1.2.2
nibabel~=4.0.1